requirements.txt
python-telegram-bot==20.7
SQLAlchemy==2.0.25
python-dotenv
